xattrs: handle both hash forms in reiserfs_check_xattr
It turns out that the csum_partial used on x86_64 is essentially the only
implementation that is 32-bit and it's intended to be folded.  Other
implementations use a 16-bit version.  This library includes the 16-bit
version so that any validation of extended attribute checksums would
fail.  The fix is to convert the 32-bit version to the 16-bit version
if the 16-bit version fails before trying again.

This affects moving file systems between architectures that define
csum_partial differently.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>